Move connection utilities to utils module. Rely on override options in test.
